initial working rust build, linker broken
we had to add some compiler directives to disable the stdlib (can't
run on bare metal with external dependencies), which led to some
other problems. we needed a global handler for `panic`s, in case of
runtime errors, as well as a new entry point for the binary.
normally the entry point is a start function called from a C entry
point, but without bundling the C stdlib there's no "normal" entry
point to call.

at this point, we still don't have a functioning binary entry, just a
rust `_start` function that's externalized. still need to hook that
up to something.