add custom build target,use xbuild to compile core
not a ton of code here, but a fair amount got done. set up the target
image to build for bare metal instead of building for linux, used
xbuild so that we recompile the core libraries for this new environment
instead of using the linux-compiled binaries, and updated github
build to use the new target. also set up a cargo config so we don't
need to specify the target on every build command