make line buffering work properly
lots of fun stuff in this one. implements newline functionality, so
the buffer knows when it's at the edge of the screen, moves the rest of
the lines up, clears the bottom line, and starts inputting at the start
of the line again. also implements the std::fmt::Write trait so we can
use rust's standard write! macro to output text to our screen.