static writer instance rather than generated
this is mostly copypasta. this statically allocates the Writer rather
than having to build your own writer, which guarantees that we can't
have multiple writers overwriting each others' spaces in memory. instead
we have a statically allocated buffer in the writer and a single
interface to the VGA output. this involves some third-party packages,
a lazy-static package to allow the members of a static object to be
non-static, and a custom version of Mutex that works when we don't have
a concept of threads, which is the case here when we're running on bare
metal.