create custom test runner and conditionally run
because we're not including the std library, it means that we also don't
get access to the built-in test runner :roll_eyes: fortunately, we can
add some compiler directives to use a custom test runner, define a
custom test-main function, and then conditionally execute it depending
on whether or not we're in a test build or not. i'm going to want to do
a little metaprogramming to make the progress report print
automatically, maybe print out some stats (runtime, %success, etc), and
i'm not sure what else.