add serial interface to print test output
when our tests run in the virtual OS, they print to the screen really
fast and then the program exits, which isn't ideal. there's no way to
see your test results. additionally, ideally we wouldn't launch the qemu
env at all, there's nothing happening there that we actually need to see
during a test. instead we should print output to the console. since
we're already tied to qemu for emulation of the OS, we can use the fact
that QEMU implements serial ports and write our output to the serial
bus instead of the VGA output